win state working
